Feat/#12 예산 수정 api

+++ src/main/java/com/goormthon/whattoeat/repository/BudgetRepository.java
@@ -3,5 +3,5 @@ package com.goormthon.whattoeat.repository;
 import com.goormthon.whattoeat.domain.Budget;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
 }
